refactor(lexer): split `carve` pipeline stage into multiple files (#26416)
Pure refactor. No code changes, just moving code around.

Split `carve` pipeline stage into multiple files.

Unlike #26390, here there's little platform-specific code to split up. Aim is just to separate out code which is specific to JS / JSX, and split JSX code up into separate sub-systems.

Additionally, apply the same kinds of stylistic changes as #26390.
