+++ crates/oxc_lexer/src/pipeline/carve/jsx/common.rs
@@ -1,0 +1,4 @@
+#[inline(always)]
+pub(super) unsafe fn wordbit(word: *const u64, p: usize) -> bool {
+    (*word.add(p >> 6) >> (p & 63)) & 1 != 0
+}
